Halo creates a buffer around pixel bigger than 0

Halo creates a buffer around pixel bigger than 0

Usage

Halo(im, width = 2, halo.only = TRUE, kernel = "circle")

Arguments

im

segmented raster

width

buffer around roots in px, the rhizosphere extent. Exudation diffusion distance is typically 2mm (1-12mm) (Finzi et al. 2015, https://doi.org/10.1111/gcb.12816), but higher values have been suggested. At 300 dpi, 1mm = 11.8px

halo.only

set TRUE if only the buffer around roots should be returned (the rhizosphere only)

kernel

the thickening shape. Available options are: "circle" & "diamond"

Value

SpatRaster

Examples

library(terra)

data(seg_Oulanka2023_Session01_T067)
img = terra::rast(seg_Oulanka2023_Session01_T067)
Halo(im = img, width = 2, halo.only = TRUE)
#> class       : SpatRaster 
#> dimensions  : 1144, 4900, 3  (nrow, ncol, nlyr)
#> resolution  : 0.0002040816, 0.0008741259  (x, y)
#> extent      : 0, 1, 0, 1  (xmin, xmax, ymin, ymax)
#> coord. ref. :  
#> source(s)   : memory
#> names       : lyr1, lyr2, lyr3 
#> min values  :    0,    0,    0 
#> max values  :    3,    3,    3