Estimate the position of the soil surface by tape presence

Estimate the position of the soil surface by tape presence

Usage

SoilSurfE(
  img,
  search.area = 0.45,
  tape.tresh = 0.33,
  dpi = 300,
  nclasses = 3,
  inverse = FALSE,
  tape.overlap = 0.5,
  tape.brightness = 0.6,
  extra.rows = 100,
  tape.quantile = 0.98
)

Arguments

img

raster,filename, or array input

search.area

ratio of image which is used to look for tape cover. Speeds up computation.

tape.tresh

ratio of how much of the tube rotation needs to covered in tape

dpi

image resolution

nclasses

number of clusters to discern tape, roots, background. 3 for rgb images. 2 can be appropriate for segmented 0-1 images.

tape.overlap

assumes a safety margin on the tape. The soil surface will be shifted by this amount in cm

tape.brightness

used for clustering. Tape appears bright e.g., 0.66

extra.rows

In case no tape is present. Best leave unchanged - some extra.rows are recommended and will be subtracted from the output anyway.

tape.quantile

aligns extra.rows brightness with the tape. The default uses Silver Tape as reference.

Value

data.frame with tape end and soil surface estimation in rows

Examples

img = seg_Oulanka2023_Session01_T067
Soil0Estimates = SoilSurfE(img)