Extracts Root Decay, New Root Production, and No-Change Roots. Relies on 'RootDetector'

Extracts Root Decay, New Root Production, and No-Change Roots. Relies on 'RootDetector'

Usage

Turnover.DPC(
  img,
  product.layer = 2,
  decay.layer = 1,
  blur.capture = 0.95,
  im.return = FALSE,
  include.virtualroots = FALSE
)

Arguments

img

image in the 'RootDetector' format - one layer for production, one layer for decay, one layer for stagnation

product.layer

layer indicating production

decay.layer

layer indicating decay & tape

blur.capture

pixel are included if: value >= max value * blur.capture. Ensures that attenuated pixels (as a result of blurring or resizing) are also included

im.return

return images instead of values?

include.virtualroots

should all roots which were present at some point in one of the two time steps be considered? Consider decay in production ratio, and production in decay ratio?

Value

either pixel sums if im.return = F, or individual layers corresponding to tape, production, decay, and no change

Examples

data(TurnoverDPC_data)
img = terra::rast(TurnoverDPC_data)
DPCs = Turnover.DPC(img = img, im.return = FALSE)