Global root production and root turnover from temporal comparison

Global root production and root turnover from temporal comparison

Usage

Turnover.TC(im.t1, im.t2, method = "kimura", unit = "cm", dpi = 300)

Arguments

im.t1

brick raster timepoint 1

im.t2

brick raster timepoint 2

method

choose between "kimura" & "rootpx". Adjust the input image accordingly

unit

unit of root length. only applies if method = "kimura"

dpi

image resolution. only applies if method = "kimura"

Value

standing roots at the start, standing roots at the end, root production,

Details

Chosing kimura as method will return root length with default settings from RootLength()

Examples

data(skl_Oulanka2023_Session01_T067)
data(skl_Oulanka2023_Session03_T067)
time1 = terra::rast(skl_Oulanka2023_Session01_T067)
time2 = terra::rast(skl_Oulanka2023_Session03_T067)
turnover.values = Turnover.TC(im.t1 = time1, im.t2 = time2, method= "kimura")