Finite sample/Large sample Wald test for differential phase.

Test differential phase of circadian curve fitting using Wald test

Usage

WaldTest_diff_phase(tt1, yy1, tt2, yy2, period = 24, FN = TRUE)

Arguments

tt1

Time vector of condition 1

yy1

Expression vector of condition 1

tt2

Time vector of condition 2

yy2

Expression vector of condition 2

period

Period of the since curve. Default is 24.

FN

Type of test to use, TRUE = "FN" (finite) or FALSE = "LS" (large samples). Default is finite sample.

Value

A list, see details below. Formula 1: \(yy = amp \times sin(2\pi/period \times (phase + tt)) + offset\) Formula 2: \(yy = A \times sin(2\pi/period \times tt) + B * cos(2*pi/period * tt) + offset\)

phase_1

Phase estimate of the 1st data, phase is restricted in (0, period)

phase_2

Phase estimate of the 2nd data, phase is restricted in (0, period)

phase_c

Phase estimate pooling all data together, phase is restricted in (0, period)

df

Degree of freedom for the Wald test

stat

Wald statistics

pvalue

P-value from the Wald test

Author

Caleb

Examples

set.seed(32608)
n <- 50
Period <- 24
tt1 <- runif(n,0,Period)
Amp1 <- 2
Phase1 <- 6
Offset1 <- 3
yy1 <- Amp1 * sin(2*pi/Period * (tt1 + Phase1)) + Offset1 + rnorm(n,0,1)
tt2 <- runif(n,0,Period)
Amp2 <- 3
Phase2 <- 5
Offset2 <- 2
yy2 <- Amp2 * sin(2*pi/Period * (tt2 + Phase2)) + Offset2 + rnorm(n,0,1)
WaldTest_diff_phase(tt1, yy1, tt2, yy2)
#> $phase_1
#> [1] 6.536394
#> 
#> $phase_2
#> [1] 5.399322
#> 
#> $phase_c
#> [1] 5.797734
#> 
#> $stat
#> [1] 8.523932
#> 
#> $pvalue
#> [1] 0.005681205
#>