Fisher information matrix when two conditions exist

Obtain the Fisher information matrix when two conditions exist

Usage

fisherInformation2(beta, tt1, yy1, tt2, yy2, period = 24)

Arguments

beta

parameter vector of 8 with the following order: amp_1, phase_1, offset_1, theta_1, amp_2, phase_2, offset_2, theta_2

tt1

time vector of condition 1

yy1

expression vector of condition 1

tt2

time vector of condition 2

yy2

expression vector of condition 2

period

Period of the since curve. Default is 24.

Value

The Fisher information matrix, this is a 8*8 matrix, with the same order as the input beta parameter.

Author

Caleb

Examples

set.seed(32608)
n <- 50
Period <- 24
tt1 <- runif(n,0,Period)
Amp1 <- 2
Phase1 <- 6
Offset1 <- 3
yy1 <- Amp1 * sin(2*pi/Period * (tt1 + Phase1)) + Offset1 + rnorm(n,0,1)
tt2 <- runif(n,0,Period)
Amp2 <- 3
Phase2 <- 15
Offset2 <- 2
yy2 <- Amp2 * sin(2*pi/Period * (tt2 + Phase2)) + Offset2 + rnorm(n,0,1)
beta <- c(Amp1,Phase1,Offset1,1,Amp2,Phase2,Offset2,2)
fisherInformation2(beta, tt1, yy1, tt2, yy2)
#>           [,1]      [,2]       [,3]       [,4]       [,5]      [,6]       [,7]
#> [1,] 25.456103  1.484273 -3.6592758  1.1964346  0.0000000  0.000000   0.000000
#> [2,]  1.484273  6.564844 -3.0547070 -3.0476574  0.0000000  0.000000   0.000000
#> [3,] -3.659276 -3.054707 50.0000000 -0.2308641  0.0000000  0.000000   0.000000
#> [4,]  1.196435 -3.047657 -0.2308641 25.0000000  0.0000000  0.000000   0.000000
#> [5,]  0.000000  0.000000  0.0000000  0.0000000 46.6709485  2.940024   2.143862
#> [6,]  0.000000  0.000000  0.0000000  0.0000000  2.9400244 32.965075  -6.824443
#> [7,]  0.000000  0.000000  0.0000000  0.0000000  2.1438620 -6.824443 100.000000
#> [8,]  0.000000  0.000000  0.0000000  0.0000000 -0.1678738  7.157126  12.068982
#>            [,8]
#> [1,]  0.0000000
#> [2,]  0.0000000
#> [3,]  0.0000000
#> [4,]  0.0000000
#> [5,] -0.1678738
#> [6,]  7.1571265
#> [7,] 12.0689824
#> [8,]  6.2500000