An original Minirhizotron Root Scan from Timepoint 2

This is a blended image of multiple scans stitched together. The scan originates from a sedge fen in northern Finland in late October. Columns correspond to tube length and rows to tube rotation

Usage

data(rgb_Oulanka2023_Session03_T067)

Format

a RasterBrick with 4900 columns x 1161 rows x 3 layer

Source

Images by J.Cunow

Author

Johannes Cunow johannes.cunow@gmail.com